Fix bigInteger references

+++ database/migrations/2015_10_24_000000_create_txos_table.php
@@ -18,7 +18,7 @@ class CreateTxosTable extends Migration {
 
             $table->char('txid', 64)->index();
             $table->integer('n')->unsigned();
-            $table->integer('amount')->unsigned()->default(0);
+            $table->bigInteger('amount')->unsigned()->default(0);
             $table->text('script')->default('');
 
             $table->integer('type')->unsigned()->default(1);
